Here is version two of the chainmail tiling
Changed up the code to instead do a semitransparent overlay of the original picture, rather than finding the color each one should be, since that is less useful than I thought. Also added line numbers as QOL update. This isn't an edit of the previous ipynb notebook on here since I'm too lazy to set up a github for this on my local machine, since this is probably the last edit I make to this